:sparkles: #22 Calculate fibonacci to log progress
Add a new fibonacci calculator that will allow to print a log in a progressive manner depending on the size of the sample. this ways is easier to know how many records have been processed. this algorithm will calculate and cache all the existing values making it very optimal. This can be used in other scenarios as well